requirements.txt
torch >= 1.3.0
numpy >= 1.16.4
opencv-python >= 4.1.0
PyTurboJPEG
tensorflow
keras~=2.12.0